+++ Package.swift
@@ -18,6 +18,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/swiftlang/swift-syntax.git", from: "600.0.0"),
+        .package(url: "https://github.com/apple/swift-docc-plugin.git", from: "1.4.0"),
     ],
     targets: [
         // MARK: - Macro Implementation